+++ doc/whats-new.rst
@@ -27,6 +27,8 @@ Bug fixes
 
 - Fix a bug in lazily-indexing netCDF array. (:issue:`1688`)
   By `Keisuke Fujii <https://github.com/fujiisoup>`_
+- (Internal bug) MemoryCachedArray now supports the orthogonal indexing.
+  Also made some internal cleanups around array wrappers (:issue:`1429`).
 
 - Fix two bugs that were preventing dask arrays from being specified as
   coordinates in the DataArray constructor (:issue:`1684`).